extract headers
