+++ resources/js/Components/Tables/CustomerTableWithActions.vue
@@ -5,45 +5,9 @@
                 
                 <div
                     class="grid items-center justify-between grid-flow-col p-4 space-y-3 border-b md:flex-row md:space-y-0 md:space-x-4 dark:border-gray-700">
-                    <div class="flex flex-row flex-1">
-                        <!-- SEARCH -->
-                        <form class="flex-1 w-full md:max-w-sm md:mr-4">
-                            <label for="search-orders"
-                                class="text-sm font-medium text-gray-900 sr-only dark:text-white">{{ $t('labels.search') }}</label>
-                            <div class="relative grid grid-flow-col">
-                                <input type="search" id="search-orders"
-                                    class="block w-full p-2 pl-5 pr-4 text-sm text-gray-900 border border-gray-300 rounded-lg rounded-e-none border-e-0 min-w-40 bg-gray-50 focus:border-gray-400 focus:ring-0 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400"
-                                    :placeholder="$t('labels.search')">
-                                <button id="limitSearchFilterDropdownButton" data-dropdown-toggle="limitSearchFilterDropdown"
-                                        class="flex items-center justify-center w-full px-2 py-2 text-sm font-medium text-gray-900 border border-gray-300 bg-gray-50 md:w-auto focus:outline-none hover:bg-gray-100 hover:text-primary-700 focus:z-10 focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:text-white dark:hover:bg-gray-700"
-                                    type="button">
-                                        <el-icon><Filter /></el-icon>
-                                        <!-- {{ $t('labels.limit-search') }} -->
-                                        <el-icon size="24" class="pl-2"><ArrowDown /></el-icon>
-                                </button>
-                                <button @click.prevent="reorder(_headers)" type="submit"
-                                    class="relative top-0 bottom-0 right-0 px-4 py-2 text-sm font-medium text-white rounded-r-lg bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:outline-none focus:ring-primary-300 dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800">{{ $t('labels.search') }}</button>
-                            </div>
-                        </form>
-                        <!-- SEARCH End -->
-                        <!-- Search in Fields -->
-                        <div id="limitSearchFilterDropdown"
-                            class="hidden w-auto p-3 bg-white rounded-lg shadow dark:bg-gray-700">
-                            <h6 class="mb-3 text-sm font-medium text-gray-900 dark:text-white">{{ $t('labels.select-fields')}}</h6>
-                            <ul class="space-y-2 text-sm" aria-labelledby="limitSearchFilterDropdownButton">
-                                <li v-for="head in _headers" class="flex items-center">
+                    
+                    <FilteredSearch :searchTerm="data" searchAt="customers.index" :headers="_headers" />
 
-                                    <input v-if="head.searchable == true" :id="'search-label-' + head.key" type="checkbox"
-                                        :value="head.key"
-                                        class="w-4 h-4 bg-gray-100 border-gray-300 rounded text-primary-600 focus:ring-primary-500 dark:focus:ring-primary-600 dark:ring-offset-gray-700 focus:ring-2 dark:bg-gray-600 dark:border-gray-500">
-                                    <label v-if="head.searchable == true" :for="'search-label-' + head.key"
-                                        class="block w-full ml-2 text-sm font-medium text-gray-900 cursor-pointer hover:text-corporate-700 dark:text-gray-100">{{ $t(head.title) }}</label>
-                                </li>
-                            </ul>
-                        </div>
-                        <!-- Search in Fields end -->
-                        
-                    </div>
                     <div class="grid grid-flow-col gap-4">
                         <!-- CREATE ORDER BUTTON -->
                         <button type="button"
@@ -85,7 +49,7 @@
                         <thead class="text-xs uppercase bg-gray-50 dark:bg-gray-700">
                             <tr>
 
-                                <th v-for="header in headers" scope="col" class="px-4 py-3">{{ $t(header.text) }}
+                                <th v-for="header in _headers" scope="col" class="px-4 py-3">{{ $t(header.text) }}
                                     <svg v-if="header.orderable == true" class="inline-block w-4 h-4 ml-1"
                                         fill="currentColor" viewbox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"
                                         aria-hidden="true">
@@ -102,7 +66,7 @@
                             <tr v-for="entry in data " :key="entry.id"
                                 class="transition border-b cursor-pointer dark:border-gray-700 hover:bg-gray-200 dark:hover:bg-gray-700">
 
-                                <td v-for="head in headers" class="px-4 py-3">
+                                <td v-for="head in _headers" class="px-4 py-3">
                                     <span v-if="head.key == 'private_customer' && entry['private_customer'] == true" class="bg-green-100 text-green-800 text-xs font-medium me-2 px-2.5 py-0.5 rounded-full dark:bg-green-900 dark:text-green-300">{{ $t('labels.customer-private') }}</span>
                                     <span v-else-if="head.key == 'private_customer' && entry['private_customer'] == false" class="bg-purple-100 text-purple-800 text-xs font-medium me-2 px-2.5 py-0.5 rounded-full dark:bg-purple-900 dark:text-purple-300">{{ $t('labels.customer-business') }}</span>
                                     <span v-else>{{ entry[head.key] }}</span>
@@ -147,27 +111,13 @@
 </template>
 
 <script setup>
-import { DArrowLeft, ArrowLeft, DArrowRight, ArrowRight, Plus, View, ArrowDown, Filter } from '@element-plus/icons-vue';
+import { Plus, View, ArrowDown, Filter } from '@element-plus/icons-vue';
 import { router } from '@inertiajs/vue3';
 import { initFlowbite } from 'flowbite';
 import { onMounted, ref} from 'vue';
-defineProps({
-    getData: {
-        type: Function,
-        required: true,
-    },
-    paginationData: {
-        type: Object,
-        required: true,
-    },
-    totalResults: {
-        type: Number,
-        required: false,
-    },
-    title: {
-        type: String,
-        required: false,
-    },
+import FilteredSearch from '@/Components/Inputs/FilteredSearch.vue';
+
+const props = defineProps({
     headers: {
         type: Object,
         required: false,
@@ -182,24 +132,9 @@ defineProps({
     },
 })
 
-
+const _headers = props.headers;
 
 onMounted(() => {
     initFlowbite();
 })
-
-const handleShow = (entry_id) => {
-    alert(`handleShow for item: ${entry_id}`)
-}
-const handleEdit = (entry_id) => {
-    alert(`handleEdit for item: ${entry_id}`)
-}
-const handleDelete = (entry_id) => {
-    alert(`handleDelete for item: ${entry_id}`)
-}
-
-</script>
-
-<script>
-
 </script>
