+++ resources/js/Pages/Customers/Index.vue
@@ -6,7 +6,7 @@
         :actions="['show', 'edit']"
         :title="$t('labels.customers')"
         :data="fe_data.customers"
-        :headers="headers"
+        :headers="defaultHeaders"
     />
     <Pagination :links="fe_data.pagination" />
 </template>
@@ -21,36 +21,8 @@ import { useCustomerStore } from '@/Stores/customerStore';
 import CustomerTableWithActions from '@/Components/Tables/CustomerTableWithActions.vue';
 import Pagination from "@/Components/Pagination/Pagination.vue";
 
-const headers = [
-    {
-        key: "id",
-        text: 'labels.customer-id',
-        searchable: true,
-    },
-    {
-        key: "first_name",
-        text: 'labels.first-name',
-        searchable: true,
-        orderable: true
-    },
-    {
-        key: "last_name",
-        text: 'labels.last-name',
-        searchable: true,
-        orderable: true
-    },
-    {
-        key: "company_name",
-        text: 'labels.company-name',
-        searchable: true,
-    },
-    {
-        key: "private_customer",
-        text: 'labels.business-customer',
-        searchable: true,
-        orderable: true
-    }
-]
+import headers from "@/config/Tables/customerHeaders";
+const defaultHeaders = headers;
 
 let props = defineProps( 
     {
