+++ resources/js/Pages/Forwarders/Index.vue
@@ -3,6 +3,9 @@
 </template>
 <script setup>
 
+import headers from "@/config/Tables/forwarderHeaders";
+const defaultHeaders = headers;
+
 
 const props = defineProps({
     records: {
