+++ resources/js/config/Tables/customerHeaders.js
@@ -1,4 +1,38 @@
-const headers = [];
+const headers = [
+    {
+        show: true,
+        key: "id",
+        text: 'labels.customer-id',
+        searchable: true,
+    },
+    {
+        show: true,
+        key: "first_name",
+        text: 'labels.first-name',
+        searchable: true,
+        orderable: true
+    },
+    {
+        show: true,
+        key: "last_name",
+        text: 'labels.last-name',
+        searchable: true,
+        orderable: true
+    },
+    {
+        show: true,
+        key: "company_name",
+        text: 'labels.company-name',
+        searchable: true,
+    },
+    {
+        show: true,
+        key: "private_customer",
+        text: 'labels.business-customer',
+        searchable: true,
+        orderable: true
+    }
+]
 
 
 export default headers;
