+++ resources/js/config/Tables/orderHeaders.js
@@ -1,4 +1,152 @@
-const headers = [];
+const headers = [
+    {
+        show: true, 
+        key: "id", 
+        title: "labels.id", 
+        sortable: true, 
+        filterable: true,
+        searchable: true,
+        display_order: 1
+    },{
+        show: true, 
+        key: "forwarder", 
+        title: "labels.forwarder", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 1
+    },{
+        show: true, 
+        key: "type_of_transport", 
+        title: "labels.type_of_transport", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 2
+    },{
+        show: true, 
+        key: "origin", 
+        title: "labels.origin", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 3
+    },{
+        show: true, 
+        key: "status", 
+        title: "labels.status", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 4
+    },{
+        show: true, 
+        key: "customer_reference", 
+        title: "labels.customer_reference", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 5
+    },{
+        show: true, 
+        key: "provision", 
+        title: "labels.provision", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 6
+    },
+    // {
+    //     show: true, 
+    //     key: "order_edited_events", 
+    //     title: "labels.order_edited_events", 
+    //     sortable: true, 
+    //     filterable: true, 
+    //     searchable: true,
+    //     display_order: 7
+    // },
+    {
+        show: false, 
+        key: "currency", 
+        title: "labels.currency", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "order_date", 
+        title: "labels.order_date", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "purchase_price", 
+        title: "labels.purchase_price", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "month_and_year", 
+        title: "labels.month_and_year", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "created_at", 
+        title: "labels.created_at", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "updated_at", 
+        title: "labels.updated_at", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "payment_method", 
+        title: "labels.payment_method", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "easy_bill_customer_id", 
+        title: "labels.easy_bill_customer_id", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "last_update", 
+        title: "labels.last_update", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    },{
+        show: false, 
+        key: "last_editor", 
+        title: "labels.last_editor", 
+        sortable: true, 
+        filterable: true, 
+        searchable: true,
+        display_order: 7
+    }
+];
 
 
 export default headers;
